+++ frontend/src/components/ChatSettings/ChatSettings.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import {
-	SettingsContainer,
+	MainContentWrapper,
 	FooterWrapper,
 	HeaderWrapper,
-} from "../Settings/SettingsElements";
-import { MainContentWrapper } from "../Utils/Containers/Containers";
+	MainViewContainer,
+} from "../Utils/Containers/Containers";
 import { User, ChatData, ToSend } from "../../Types/Types";
 import { fetchData, postData } from "../../API/API";
 import { Text, Header } from "../Utils/Text/Text";
@@ -121,8 +121,8 @@ const ChatSettings = (): JSX.Element => {
 
 					{myRole != roleLevel.user ? (
 						<>
-							<Text>Add Users</Text>
 							<UserWrapper>
+								<Text>Add Users</Text>
 								<AddUserInput
 									removeOnEnter={true}
 									placeholder={"Enter username"}
@@ -131,7 +131,6 @@ const ChatSettings = (): JSX.Element => {
 									}}
 								/>
 								<br />
-								<hr />
 								{usersToAdd ? listAddedUsers : ""}
 							</UserWrapper>
 						</>
@@ -187,9 +186,9 @@ const ChatSettings = (): JSX.Element => {
 	};
 
 	return (
-		<SettingsContainer>
+		<MainViewContainer>
 			{chatData ? displayChat() : "loading"}
-		</SettingsContainer>
+		</MainViewContainer>
 	);
 };
