+++ frontend/src/components/ConnectionForm/ConnectionFormElements.ts
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { Link as LinkR } from "react-router-dom";
 
 export const FormContainer = styled.div`
 	display: flex;
@@ -18,41 +17,6 @@ export const FormContainer = styled.div`
 export const Form = styled.form`
 	padding: 10px;
 	text-align: center;
-`;
-
-export const Button = styled.button`
-	margin: 10px;
-	background-color: #3f3fff;
-	color: white;
-	padding: 12px 20px;
-	border: none;
-	border-radius: 4px;
-	cursor: pointer;
-
-	&:hover {
-		background-color: #abc;
-	}
-`;
-
-export const Label = styled.label`
-	color: #fff;
-	padding: 12px 12px 12px 0;
-	display: inline-block;
-`;
-
-export const Btn = styled(LinkR)`
-	height: 50px;
-	width: 50px;
-	text-align: center;
-	background-color: #444;
-	color: white;
-	border-radius: 50px;
-	border: 2px solid #3f3fff;
-	cursor: pointer;
-
-	&:hover {
-		background-color: #abc;
-	}
 `;
 
 export const RegistrationContainer = styled.div`
