--- frontend/src/components/Settings/SettingsElements.ts
@@ -1,30 +1,0 @@
-import styled from "styled-components";
-
-export const SettingsContainer = styled.div`
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-	border: 5px white solid;
-	border-radius: 5px;
-	background-color: black;
-	border-radius: 7px;
-	padding: 20px;
-	height: -webkit-fill-available;
-`;
-
-export const UsersContainer = styled.div`
-	background-color: #ad1b1b;
-	border-radius: 7px;
-	padding: 20px;
-	height: 100px;
-	width: 400px;
-`;
-
-export const HeaderWrapper = styled.div``;
-
-export const FooterWrapper = styled.div`
-	height: 110px;
-	padding: 20px;
-	display: flex;
-	justify-content: right;
-`;
