+++ frontend/src/components/Utils/Containers/Containers.ts
@@ -14,3 +14,30 @@ export const MainContentWrapper = styled.div`
 	height: 80vh;
 	overflow-y: auto;
 `;
+
+export const HeaderWrapper = styled.div``;
+
+export const FooterWrapper = styled.div`
+	height: 110px;
+	padding: 20px;
+	display: flex;
+	justify-content: right;
+`;
+
+export const MainViewContainer = styled.div`
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
+	border: 5px white solid;
+	border-radius: 5px;
+	background-color: black;
+	border-radius: 7px;
+	padding: 20px;
+	height: -webkit-fill-available;
+`;
+
+export const AdminContainer = styled.div`
+	background-color: white;
+	border-radius: 7px;
+	padding: 20px;
+`;
