+++ frontend/src/components/Utils/Label/Label.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Label = styled.label`
+	color: #fff;
+	padding: 12px 12px 12px 0;
+	display: inline-block;
+`;
